+++ jgclark.Dashboard/src/react/components/Dashboard.jsx
@@ -54,7 +54,8 @@ const Dashboard = ({ pluginData }: Props): React$Node => {
   useWatchForResizes(sendActionToPlugin)
   // 5s hack timer to work around cache not being reliable (only runs for users, not DEVs)
   // FIXME:
-  const { refreshTimer } = useRefreshTimer({ maxDelay: 5000, enabled: logSettings._logLevel !== "DEV" })
+  const shortDelayTimerIsOn = true // logSettings._logLevel !== "DEV"
+  const { refreshTimer } = useRefreshTimer({ maxDelay: 5000, enabled: shortDelayTimerIsOn })
 
   //----------------------------------------------------------------------
   // Refs
@@ -156,7 +157,8 @@ const Dashboard = ({ pluginData }: Props): React$Node => {
     if (reactSettings?.lastChange && typeof reactSettings.lastChange === 'string' && reactSettings.lastChange.length > 0 && reactSettings.lastChange[0] !== '_') {
       const trimmedReactSettings = { ...reactSettings, lastChange: '_Saving', dialogData: { isOpen: false, isTask: true, details: {} } }
       const strReactSettings = JSON.stringify(trimmedReactSettings)
-      sendActionToPlugin('reactSettingsChanged', { actionType: 'reactSettingsChanged', dashboardSettings: strReactSettings }, 'Dashboard reactSettings updated', true)
+      //FIXME: (from @dwertheimer) this whole effect can probably be deleted because we are not persisting reactSettings to the plugin anymore
+      // sendActionToPlugin('reactSettingsChanged', { actionType: 'reactSettingsChanged', dashboardSettings: strReactSettings }, 'Dashboard reactSettings updated', true)
     }
   }, [reactSettings])
 
@@ -171,11 +173,10 @@ const Dashboard = ({ pluginData }: Props): React$Node => {
       logDebug('Dashboard', `Watcher for dashboardSettings changes. Shared settings updated: "${dashboardSettings.lastChange}" sending to plugin to be saved`, dashboardSettings)
       const sharedSets = { ...dashboardSettings }
       if (settingsNeedFirstTimeMigration) { 
-        sharedSets.settingsMigrated = true 
+        sharedSets.settingsMigrated = true //FIXME: (@jgclark) these two fields should be added to your settings typedef
         sharedSets.lastChange = '_Saving first time migration'
       }
-      const strSharedSetings = JSON.stringify(sharedSets)
-      sendActionToPlugin('dashboardSettingsChanged', { actionType: 'dashboardSettingsChanged', settings: strSharedSetings, logMessage: `Settings needed migrating` }, 'Dashboard dashboardSettings updated', true)
+      sendActionToPlugin('dashboardSettingsChanged', { actionType: 'dashboardSettingsChanged', settings: sharedSets, logMessage: `Settings needed migrating` }, 'Dashboard dashboardSettings updated', true)
     } else if (dashboardSettings && Object.keys(dashboardSettings).length > 0) {
       // logDebug('Dashboard', `Watcher for dashboardSettings changes. Shared settings updated: ${JSON.stringify(dashboardSettings,null,2)}`,dashboardSettings)
     }
@@ -272,7 +273,7 @@ const Dashboard = ({ pluginData }: Props): React$Node => {
   if (sections.length === 0) {
     return <div className="dashboard">Error: No Sections to display ...</div>
   }
-  const autoUpdateEnabled = parseInt(dashboardSettings?.autoUpdateAfterIdleTime || "0") > 0
+  const autoUpdateEnabled = parseInt(dashboardSettings?.autoUpdateAfterIdleTime || "0") > 0 //FIXME: (@jgclark) needs adding to the typedef
   //autoRefresh
   // logDebug('Dashboard', `setting typeof ${typeof dashboardSettings?.autoUpdateAfterIdleTime} ${dashboardSettings?.autoUpdateAfterIdleTime ?? '-'} / ${parseInt(dashboardSettings?.autoUpdateAfterIdleTime || "0")} / ${String(autoUpdateEnabled)}`)
   return (