+++ jgclark.Dashboard/src/react/components/Header/featureFlagItems.js
@@ -5,9 +5,9 @@ import type { TDropdownItem, TDashboardConfig } from "../../../types.js"
 
 
 const featureFlagSettingDefs = [
-    { key: 'FFlag_ForceInitialLoadForBrowserDebugging', label: 'Force Full Initial Load', tooltip: 'Rather than incremental section loading, force full initial load. Mostly useful for testing full data in a browser.' },
-    { key: 'FFlag_LimitOverdues', label: 'Limit Overdues to Last 2w', tooltip: 'Pull overdues from last 2 weeks only. Requires a refresh after setting.' },
-    { key: 'FFlag_HardRefreshButton', label: 'Show Hard Refresh Button', tooltip: 'Show button that does a full window reload with changed React components and data' },
+    { key: 'FFlag_ForceInitialLoadForBrowserDebugging', label: 'Force Full Initial Load', description: 'Rather than incremental section loading, force full initial load. Mostly useful for testing full data in a browser.' },
+    { key: 'FFlag_LimitOverdues', label: 'Limit Overdues to Last 2w', description: 'Pull overdues from last 2 weeks only. Requires a refresh after setting.' },
+    { key: 'FFlag_HardRefreshButton', label: 'Show Hard Refresh Button', description: 'Show button that does a full window reload with changed React components and data' },
 ]
 
 export const createFeatureFlagItems = (dashboardSettings: TDashboardConfig): Array<TDropdownItem> => {
@@ -16,6 +16,6 @@ export const createFeatureFlagItems = (dashboardSettings: TDashboardConfig): Arr
         key: setting.key,
         type: 'switch',
         checked: (typeof dashboardSettings !== undefined && dashboardSettings[setting.key]) ?? false,
-        tooltip: setting.tooltip,
+        description: setting.description,
     }))
 }