+++ jgclark.Dashboard/src/react/components/WebView.jsx
@@ -27,7 +27,6 @@
 import React, { useEffect, useLayoutEffect, type Node } from 'react'
 import { type PassedData } from '../../reactMain.js'
 import { type TReactSettings, type TDropdownItem } from '../../types'
-import { parseSettings, getSettingsRedacted } from '../../shared.js'
 import { createDashboardSettingsItems } from '../../dashboardSettings'
 import { createFilterDropdownItems } from './Header/filterDropdownItems.js'
 import Dashboard from './Dashboard.jsx'
@@ -68,39 +67,14 @@ export function WebView({ data, dispatch, reactSettings, setReactSettings }: Pro
    *                             HOOKS
    ****************************************************************************************************************************/
 
-  // GENERALLY SPEAKING YOU DO NOT WANT TO USE STATE HOOKS IN THE WEBVIEW COMPONENT
-  // because the plugin may need to know what changes were made so when it updates data, it will be consistent
-  // otherwise when the plugin updates data, it will overwrite any changes made locally in the Webview
-  // instead of using hooks here, save updates to data using:
-  // dispatch('UPDATE_DATA', {...data,changesToData})
-  // this will save the data at the Root React Component level, which will give the plugin access to this data also
-  // sending this dispatch will re-render the Webview component with the new data
-
-  // @jgclark wrote: @dwertheimer: I've tried to understand this, and adapt it accordingly, but I've failed. I figured redactedSettings wasn't needed, as we have separated settings into two different objects.
-  // FIXME(@jgclark): We need to go through this and clean things up to make sure it does what we need to do
-  // The goal of this whole block is to build the [dashboardSettings] object for the React app to use for all the config/settings
-  // The issue is that the first time the React app is loaded, most all of the settings will not be set, so we need to make sure 
-  // to set them to the defaults specified for each item, which is what this block was doing
-  // and I don't think is (fully) doing anymore. I made some changes but we need to verify that it's working as expected
-  // and then clean out all the commented out code and other cruft
-  const dSettingsString = data.pluginData.dashboardSettings || "" // dashboardSettings is saved as a JSON string!!
-  const dSettings = parseSettings(dSettingsString || "{}") || {}
-  // const savedSharedSettings = parseSettings(dSettings || "{}") || {}
-  // const settingsDefaults = getSettingsDefaults(createDashboardSettingsItems(savedSharedSettings, dSettings))
+  // set up dashboardSettings state using defaults as the base and then overriding with any values from the plugin saved settings
+  const dSettings = data.pluginData.dashboardSettings || {}
   const dSettingsItems = createDashboardSettingsItems(dSettings)
-  // clo(dSettingsItems, `Webview: dSettingsItems`)
   const settingsDefaults = getSettingsDefaults(dSettingsItems)
-  // const filterSettingsDefaults = getSettingsDefaults(createFilterDropdownItems(savedSharedSettings, dSettings))
   const [sectionToggles, _otherToggles] = createFilterDropdownItems(dSettings)
-  // clo(sectionToggles, `Webview: sectionToggles`)
   const filterSettingsDefaults = getSettingsDefaults(sectionToggles)
-  // clo(filterSettingsDefaults, `Webview: filterSettingsDefaults`)
   const otherSettingsDefaults = getSettingsDefaults(_otherToggles)
-  // clo(otherSettingsDefaults, `Webview: otherSettingsDefaults`)
-
-  // const combinedSettings = { ...settingsDefaults, ...filterSettingsDefaults, ...redactedSettings, ...savedSharedSettings, lastChange: `_WebView_DefaultSettings` }
   const dashboardSettingsOrDefaults = { ...settingsDefaults, ...filterSettingsDefaults, ...otherSettingsDefaults, ...dSettings, lastChange: `_WebView_DefaultSettings` }
-  // clo(dashboardSettingsOrDefaults, `Webview: dashboardSettingsOrDefaults`)
   const [dashboardSettings, setDashboardSettings] = React.useState(dashboardSettingsOrDefaults)
 
   /****************************************************************************************************************************
@@ -148,62 +122,33 @@ export function WebView({ data, dispatch, reactSettings, setReactSettings }: Pro
    */
   /**
    * Helper function to initialize settings (used for reactSettings and dashboardSettings)
+   * TODO: (@dwertheimer): This was general purpose but after dashboardSettings refactor I'm not sure it's needed anymore
    * @param {Settings} settings
    */
-  function initializeSettings({ setter, currentSettings, settingsKey, defaultSettings, effectName }: Settings) {
-    if (!setter) return
-    const pluginSettingsValue = pluginData?.settings?.[settingsKey] || ''
-    // const settingsExist = currentSettings && Object.keys(currentSettings).length > 0
-    // if (settingsExist) {
-    //   logDebug(
-    //     `Webview`,
-    //     `${effectName} initializer running. currentSettings: ${window.webkit && JSON.stringify(
-    //       currentSettings || {},
-    //       null,
-    //       2,
-    //     )} pluginSettingsValue: ${pluginSettingsValue}`, currentSettings
-    //   )
-    // }
+  // function initializeSettings({ setter, currentSettings, settingsKey, defaultSettings, effectName }: Settings) {
+  //   if (!setter) return
+  //   const pluginSettingsPassed = pluginData.dashboardSettings || {}
 
-    let parsedSettings = defaultSettings
-    if (pluginSettingsValue) {
-      try {
-        logDebug(`Webview`, `${effectName} effect: loading initial data from pluginData.dashboardSettings.${settingsKey}`)
-        parsedSettings = parseSettings(pluginData.dashboardSettings[settingsKey])
-      } catch (error) {
-        logDebug(`Webview`, `${effectName} effect: could not parse settings. ${error} pluginData.dashboardSettings.${settingsKey}: ${pluginData.dashboardSettings[settingsKey]}`)
-      }
-    }
+  //   let parsedSettings = currentSettings
+  //   if (pluginSettingsPassed) {
+  //     try {
+  //       logDebug(`Webview`, `${effectName} effect: loading initial data from pluginData.dashboardSettings.${settingsKey}`)
+  //       // parsedSettings = parseSettings(pluginData.dashboardSettings[settingsKey])
+  //       parsedSettings = { ...defaultSettings, ...pluginSettingsPassed }
+  //     } catch (error) {
+  //       logDebug(`Webview`, `${effectName} effect: could not parse settings. ${error} pluginData.dashboardSettings.${settingsKey}: ${pluginData.dashboardSettings[settingsKey]}`)
+  //     }
+  //   }
 
-    setter((prev) => ({ ...prev, ...parsedSettings, lastChange: `_Webview_firstLoad` }))
-  }
+  //   setter((prev) => ({ ...prev, ...parsedSettings, lastChange: `_Webview_firstLoad` }))
+  // }
 
   /**
    * Set up the initial React Settings (runs only on load)
    */
   useEffect(() => {
-    initializeSettings({
-      setter: setReactSettings,
-      currentSettings: reactSettings,
-      settingsKey: 'reactSettings',
-      defaultSettings: defaultReactSettings,
-      effectName: 'setReactSettings'
-    })
+    setReactSettings((prev) => ({ ...prev, ...defaultReactSettings, lastChange: `_Webview_firstLoad` }))
   }, [setReactSettings])
-
-  /**
-   * Set up the initial Shared Settings (runs only on load)
-   * TODO: Maybe remove this...may not be necessary now that I am initializing in state
-   */
-  // useEffect(() => {
-  //   initializeSettings({
-  //     setter: setDashboardSettings,
-  //     currentSettings: dashboardSettings,
-  //     settingsKey: 'dashboardSettings',
-  //     defaultSettings: combinedSettings,
-  //     effectName: 'setDashboardSettings'
-  //   })
-  // }, [setDashboardSettings])
 
 
   /**