+++ jgclark.Dashboard/CHANGELOG.md
@@ -6,6 +6,10 @@ For more details see the [plugin's documentation](https://github.com/NotePlan/pl
 - new x-callback 'setSetting' to change a single Dashboard setting. TODO: See README for details. TODO: still need to handle show*Section settings.
 - new x-callback 'setSettings' to change multiple Dashboard settings in one call.
 - new **/Make Callback from Current Settings** command that copies the current setting as a URL or a markdown link to the Clipboard.
+- dbw: remove "tooltip" prop for featureFlags
+- dbw: remove DEV limitation on 5s refresh
+- dbw: fix edge case bug on ReactSettings and saving/reading from plugin.settings
+- dbw: fix issues with logSettings not saving/reading properly and not showing DEV FFlag menu
 
 ### Changes
 - made the Filters dropdown menu a two-column display (except on narrow screens)